OSX, C, refact: move disk functions in disk.c
